# Optimise help and usage buffer size