## Summary

<!-- Describe your changes in detail here, if it closes an open issue, include "Closes #<issue>" -->
Grow the initial buffer for the help and usage text

```plaintext
goos: darwin
goarch: arm64
pkg: github.com/FollowTheProcess/cli
cpu: Apple M4 Pro
               │ before.txt  │             after.txt              │
               │   sec/op    │   sec/op     vs base               │
ExecuteHelp-14   2.251µ ± 0%   2.072µ ± 0%  -7.95% (p=0.000 n=30)

               │  before.txt  │              after.txt               │
               │     B/op     │     B/op      vs base                │
ExecuteHelp-14   5.246Ki ± 0%   4.214Ki ± 0%  -19.68% (p=0.000 n=30)

               │ before.txt │             after.txt             │
               │ allocs/op  │ allocs/op   vs base               │
ExecuteHelp-14   75.00 ± 0%   68.00 ± 0%  -9.33% (p=0.000 n=30)

```